SIZE_T is not size_t for Borland 5.82
git-svn-id: http://svn.osdn.net/svnroot/hamigaki@144 c04fc264-1e22-0410-9499-a186082fc69e

+++ hamigaki/trunk/hamigaki/coroutine/detail/fiber.hpp
@@ -43,7 +43,9 @@ extern "C" {
 
 typedef void (__stdcall* start_routine)(void*);
 
-__declspec(dllimport) void* __stdcall CreateFiber(size_t, start_routine, void*);
+__declspec(dllimport)
+void* __stdcall CreateFiber(unsigned long, start_routine, void*);
+
 __declspec(dllimport) void __stdcall DeleteFiber(void*);
 __declspec(dllimport) void* __stdcall ConvertThreadToFiber(void*);
 __declspec(dllimport) void __stdcall SwitchToFiber(void*);
